+++ lib/ts/recipe/dashboard/utils.ts
@@ -31,6 +31,7 @@ import {
     USER_PASSWORD_API,
     USER_SESSIONS_API,
     VALIDATE_KEY_API,
+    DASHBOARD_ANALYTICS_API,
 } from "./constants";
 import {
     APIInterface,
@@ -61,7 +62,6 @@ export function validateAndNormaliseUserInput(config?: TypeInput): TypeNormalise
     };
 
     return {
-        apiKey: config === undefined ? undefined : config.apiKey,
         override,
         authMode: config !== undefined && config.apiKey ? "api-key" : "email-password",
     };
@@ -144,6 +144,10 @@ export function getApiIdIfMatched(path: NormalisedURLPath, method: HTTPMethod): 
     }
     if (path.getAsStringDangerous().endsWith(SEARCH_TAGS_API) && method === "get") {
         return SEARCH_TAGS_API;
+    }
+
+    if (path.getAsStringDangerous().endsWith(DASHBOARD_ANALYTICS_API) && method === "post") {
+        return DASHBOARD_ANALYTICS_API;
     }
 
     return undefined;